0.8.103 [2023-05-12] * feat: support of contract ledger join * feat: add as_integer() for ColumnField
